refactor: Make more consistent with other kuadrant operators
* Keep Makefile consistent with other operators
* Remove gateway prefix from commands and make targets
