+++ .github/workflows/controller-image.yaml
@@ -5,6 +5,7 @@ on:
     branches:
       - main
       - "release-*"
+      - remove_gateway_prefix
     tags:
       - "v[0-9]+.[0-9]+.[0-9]+"
 
@@ -56,7 +57,6 @@ jobs:
         with:
           push: true
           tags: ${{ env.IMG_TAGS }}
-          target: controller
 
       - name: Print Image URL
         run: |
