+++ config/manager/kustomization.yaml
@@ -1,7 +1,9 @@
 resources:
 - manager.yaml
 
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
 images:
 - name: controller
   newName: quay.io/kuadrant/multicluster-gateway-controller
-  newTag: main
+  newTag: latest
