+++ docs/experimental/submariner-poc-hub-gateway-walkthrough.md
@@ -111,7 +111,7 @@ In `T2` run the following to start the Gateway Controller:
 
 ```bash
 kind export kubeconfig --name=mgc-control-plane --kubeconfig=$(pwd)/local/kube/control-plane.yaml && export KUBECONFIG=$(pwd)/local/kube/control-plane.yaml
-make build-gateway-controller install run-gateway-controller
+make build install run
 ```
 
 ### Create a Gateway
