+++ pkg/ocm/addon/addon-manager.go
@@ -1,4 +1,4 @@
-package ocm
+package addon
 
 import (
 	"context"
