+++ pkg/ocm/addon/manifests/cluster-role.yaml
