+++ pkg/ocm/addon/manifests/kuadrant-namespace.yaml
