+++ pkg/ocm/addon/manifests/kuadrant.yaml
