+++ pkg/ocm/addon/manifests/operator-group.yaml
