+++ pkg/ocm/addon/manifests/subscription.yaml
